app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-dotenv
requests
PyJWT